Add GenreRepository
