+++ src/main/java/com/jd2/moviebase/model/AccountMovieId.java
@@ -1,0 +1,18 @@
+package com.jd2.moviebase.model;
+
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Embeddable
+public class AccountMovieId {
+
+    private Long accountId;
+    private Long movieId;
+}
